tests/wpt/web-platform-tests/css/css-shapes/shape-outside-invalid-circle-002.html

<!DOCTYPE HTML>
<html>
<head>
	<meta charset="UTF-8">
	<meta name="flags" content="dom">
    <meta name="assert" content="shape-outside declaration is invalid if cy for circle function is invalid">
	<title>CSS Shapes Test - Circle - invalid cy value</title>
	<link rel="author" title="noonnightstorm" href="mailto:noonnightstorm@gmail.com">
	<link rel="reviewer" title="Alan Stearns" href="mailto:stearns@adobe.com">
	<link rel="help" href="http://www.w3.org/TR/css-shapes-1/#supported-basic-shapes">
	<link rel="help" href="http://www.w3.org/TR/css-shapes-1/#funcdef-circle">
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>

	<style>
#shape{
	width: 300px;
	height: 300px;
	shape-outside: circle(50% at 50% 50);
}
	</style>
</head>
<body>

	<div id="log"></div>

	<div id="shape"></div>

	<script>
	var obj = document.getElementById('shape');
	var cy = getComputedStyle(obj).shapeOutside;
	test(function() {assert_equals(cy, "none", "declaration should be Invalid")});
	</script>

</body>
</html>
